moved delete validator test
